Move worker logutils to `galaxy.worker` package
